feat: auth middleware, deletion of own match only

+++ database.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"errors"
 	"scores-backend/models"
 
 	"github.com/jinzhu/gorm"
@@ -16,9 +17,10 @@ func initDb() (*gorm.DB, error) {
 	db.AutoMigrate(&models.Player{})
 	db.AutoMigrate(&models.Match{})
 	db.AutoMigrate(&models.Team{})
+	db.AutoMigrate(&models.User{})
 
 	var count int
-	db.First(&models.Player{}).Count(&count)
+	db.First(&models.User{}).Count(&count)
 
 	if count == 0 {
 		player1 := models.Player{Name: "Raphi"}
@@ -28,12 +30,19 @@ func initDb() (*gorm.DB, error) {
 		player5 := models.Player{Name: "Dominik"}
 		player6 := models.Player{Name: "Roman"}
 
-		db.Create(&player1)
-		db.Create(&player2)
-		db.Create(&player3)
-		db.Create(&player4)
-		db.Create(&player5)
-		db.Create(&player6)
+		user1 := models.User{Email: "raphi011@gmail.com", Player: player1}
+		user2 := models.User{Email: "", Player: player2}
+		user3 := models.User{Email: "", Player: player3}
+		user4 := models.User{Email: "Rb1@outlook.at", Player: player4}
+		user5 := models.User{Email: "Rieder.dominik@gmail.com", Player: player5}
+		user6 := models.User{Email: "", Player: player6}
+
+		db.Create(&user1)
+		db.Create(&user2)
+		db.Create(&user3)
+		db.Create(&user4)
+		db.Create(&user5)
+		db.Create(&user6)
 	}
 
 	return db, err
@@ -51,13 +60,14 @@ func getMatches() []models.Match {
 	return matches
 }
 
-func getMatch(id int) models.Match {
+func getMatch(id uint) models.Match {
 	var match models.Match
 	db.
 		Preload("Team1.Player1").
 		Preload("Team1.Player2").
 		Preload("Team2.Player1").
 		Preload("Team2.Player2").
+		Preload("CreatedBy").
 		First(&match, id)
 
 	return match
@@ -75,14 +85,25 @@ func getTeam(player1ID, player2ID uint) models.Team {
 	return team
 }
 
-/* func playerStatistic() {
-	db.Model(&models.Match)
-}  */
+func deleteMatch(matchID uint, userEmail string) error {
+	user := getUserByEmail(userEmail)
+	match := getMatch(matchID)
 
-func deleteMatch(matchID uint) {
-	match := models.Match{Model: gorm.Model{ID: matchID}}
+	if user.ID != match.CreatedByID {
+		return errors.New("Match was not created by you")
+	}
 
 	db.Delete(&match)
+
+	return nil
+}
+
+func getUserByEmail(email string) models.User {
+	var user models.User
+
+	db.Where(&User{Email: email}).First(&user)
+
+	return user
 }
 
 func createMatch(
@@ -91,15 +112,19 @@ func createMatch(
 	player3ID uint,
 	player4ID uint,
 	scoreTeam1 int,
-	scoreTeam2 int) (models.Match, error) {
+	scoreTeam2 int,
+	userEmail string,
+) (models.Match, error) {
+	user := getUserByEmail(userEmail)
 	team1 := getTeam(player1ID, player2ID)
 	team2 := getTeam(player3ID, player4ID)
 
 	match := models.Match{
-		Team1:      team1,
-		Team2:      team2,
-		ScoreTeam1: scoreTeam1,
-		ScoreTeam2: scoreTeam2,
+		Team1:       team1,
+		Team2:       team2,
+		ScoreTeam1:  scoreTeam1,
+		ScoreTeam2:  scoreTeam2,
+		CreatedByID: user.ID,
 	}
 
 	db.Create(&match)
